Implemented Ray's proposal in generic way
