+++ server/driver.go
@@ -76,7 +76,8 @@ type FileStream interface {
 
 // Settings define all the server settings
 type Settings struct {
-	Host           string // Host to receive connections on
-	Port           int    // Port to listen on
+	ListenHost     string // Host to receive connections on
+	ListenPort     int    // Port to listen on
+	PublicHost     string // Public IP to expose (only an IP address is accepted at this stage)
 	MaxConnections int    // Max number of connections to accept
 }
